Fix uint16 truncation of compressed-mesh triangle indices
bhkCompressedMeshShape data can carry more than 65,535 vertices
(SEVMageTower05.nif has 73,156). getCollCompressedMeshShapeTris
was returning indices through a uint16 buffer, silently truncating
high vertex offsets and scrambling the imported collision mesh.
Widened the triangle-index buffer to uint32 in the DLL signature
and the ctypes caller.

+++ NiflyDLL/NiflyWrapper.hpp
@@ -136,7 +136,7 @@ extern "C" NIFLY_API int getCollShapeNormals(void* nifref, int nodeIndex, float*
 extern "C" NIFLY_API int getCollPackedStripsShapeVerts(void* nifref, int dataIndex, float* buf, int buflen);
 extern "C" NIFLY_API int getCollPackedStripsShapeTris(void* nifref, int dataIndex, uint16_t* buf, int buflen);
 extern "C" NIFLY_API int getCollCompressedMeshShapeVerts(void* nifref, int dataIndex, float* buf, int buflen);
-extern "C" NIFLY_API int getCollCompressedMeshShapeTris(void* nifref, int dataIndex, uint16_t* buf, int buflen);
+extern "C" NIFLY_API int getCollCompressedMeshShapeTris(void* nifref, int dataIndex, uint32_t* buf, int buflen);
 extern "C" NIFLY_API int getCollListShapeChildren(void* nifref, int nodeIndex, uint32_t * buf, int buflen);
 extern "C" NIFLY_API void addCollListChild(void* nifref, const uint32_t id, uint32_t child_id);
 extern "C" NIFLY_API void setCollConvexTransformShapeChild(void* nifref, const uint32_t id, uint32_t child_id);
